Trait inexor_rgf_behaviour_service_api::relation_component_behaviour_manager::RelationComponentBehaviourManager

pub trait RelationComponentBehaviourManager: Send + Sync + Lifecycle {

    // Required methods
    fn add_behaviours_to_relation(&self, relation_instance: ReactiveRelation);
    fn add_behaviours_to_relation_component(
        &self,
        relation_instance: ReactiveRelation,
        component: Component
    );
    fn add_behaviour_to_relation_component(
        &self,
        relation_instance: ReactiveRelation,
        component_behaviour_ty: &ComponentBehaviourTypeId
    );
    fn remove_behaviour_from_relation(
        &self,
        relation_instance: ReactiveRelation,
        behaviour_ty: &BehaviourTypeId
    );
    fn remove_behaviours_from_relation(
        &self,
        relation_instance: ReactiveRelation
    );
    fn remove_behaviours_from_relation_component(
        &self,
        relation_instance: ReactiveRelation,
        component: Component
    );
    fn remove_behaviours_by_key(&self, edge_key: &RelationInstanceId);
    fn remove_behaviours_by_behaviour(&self, behaviour_ty: &BehaviourTypeId);
    fn has(
        &self,
        relation_instance: ReactiveRelation,
        behaviour_ty: &BehaviourTypeId
    ) -> bool;
    fn get_all(
        &self,
        relation_instance: ReactiveRelation
    ) -> Vec<BehaviourTypeId>;
    fn get_instances_by_behaviour(
        &self,
        ty: &BehaviourTypeId
    ) -> Vec<ReactiveRelation>;
    fn connect(
        &self,
        relation_instance: ReactiveRelation,
        behaviour_ty: &BehaviourTypeId
    ) -> Result<(), BehaviourTransitionError>;
    fn disconnect(
        &self,
        relation_instance: ReactiveRelation,
        behaviour_ty: &BehaviourTypeId
    ) -> Result<(), BehaviourTransitionError>;
    fn reconnect(
        &self,
        relation_instance: ReactiveRelation,
        behaviour_ty: &BehaviourTypeId
    ) -> Result<(), BehaviourTransitionError>;

}

Required Methods

fn add_behaviours_to_relation(&self, relation_instance: ReactiveRelation)

Adds new behaviours to the given relation instance.

fn add_behaviours_to_relation_component( &self, relation_instance: ReactiveRelation, component: Component )

Possibly adds new behaviour to the given relation instance’s component

fn add_behaviour_to_relation_component( &self, relation_instance: ReactiveRelation, component_behaviour_ty: &ComponentBehaviourTypeId )

Creates and adds the given behaviour to the given reactive entity instance’s component.

fn remove_behaviour_from_relation( &self, relation_instance: ReactiveRelation, behaviour_ty: &BehaviourTypeId )

Removes the given behaviour from the given reactive relation instance.

fn remove_behaviours_from_relation(&self, relation_instance: ReactiveRelation)

Removes behaviours from the given relation instance.

fn remove_behaviours_from_relation_component( &self, relation_instance: ReactiveRelation, component: Component )

Removes behaviour from the given relation instance’s component

fn remove_behaviours_by_key(&self, edge_key: &RelationInstanceId)

Removes behaviours from the given relation instance by edge key.

fn remove_behaviours_by_behaviour(&self, behaviour_ty: &BehaviourTypeId)

Removes all behaviours of the given behaviour type.

fn has( &self, relation_instance: ReactiveRelation, behaviour_ty: &BehaviourTypeId ) -> bool

Returns true, if the relation instance has the given behaviour.

fn get_all(&self, relation_instance: ReactiveRelation) -> Vec<BehaviourTypeId>

Returns the behaviours of the given relation instance.

fn get_instances_by_behaviour( &self, ty: &BehaviourTypeId ) -> Vec<ReactiveRelation>

Returns the relation instances with the given behaviour.

fn connect( &self, relation_instance: ReactiveRelation, behaviour_ty: &BehaviourTypeId ) -> Result<(), BehaviourTransitionError>

Connect

fn disconnect( &self, relation_instance: ReactiveRelation, behaviour_ty: &BehaviourTypeId ) -> Result<(), BehaviourTransitionError>

Disconnect

fn reconnect( &self, relation_instance: ReactiveRelation, behaviour_ty: &BehaviourTypeId ) -> Result<(), BehaviourTransitionError>

Reconnect